feat(main): renamed entry point
