+++ src/satclip/__init__.py
@@ -3,5 +3,5 @@ __version__ = '2025.3.2'
 
 from .location_encoder import *
 from .loss import *
-from .main import *
+from .__main__ import *
 from .model import *
